# Implement buffer pool optimizations with memory safety fixes