@robcole 2025-07-11
@gemini-code-assist I've addressed the high-severity issue with the non-existent reset_for_reuse method call in object_pool.cr. Could you please provide specific feedback on what error handling improvements are needed in tls.cr? The current implementation has try/catch blocks and proper resource cleanup. What specific improvements would you recommend?