@robcole 2025-07-11
## ✅ Resolved: High-Severity Issue Fixed

**Issue**: Non-existent method call  in 

**Resolution**: 
- Fixed in commit f6c03fb
- Disabled stream pooling due to memory safety concerns
- Converted to safe direct allocation while maintaining API compatibility
- All tests continue to pass (433 examples, 0 failures, 0 errors)
- Buffer pool performance benefits preserved (84% improvement)

**Status**: ✅ **RESOLVED** - No further action needed for this issue.

The high-severity issue identified by @gemini-code-assist has been completely addressed.