@robcole 2025-07-11
Splitting this large PR into atomic changes for better reviewability:

🔗 **PR #65: Buffer Pool Optimization** - 84% Performance Improvement
🔗 **PR #66: I/O Concurrency Optimization** - 86.6% Syscall Reduction  
🔗 **PR #67: Frame Processing Optimization** - Memory Safety & Reference Counting
🔗 **PR #68: Code Cleanup** - Remove Obsolete reset_for_reuse Methods

Each PR can now be reviewed and merged independently, making the review process more manageable and atomic.